% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{add_gitignore}
\usage{
add_gitignore(
  path = rstudioapi::selectDirectory(caption = "Select folder to add .gitignore"),
  append = NULL
)
}
\arguments{
\item{path}{String: path to add .gitignore file. If left blank, will prompt the user within R Studio}

\item{append}{Logical: if a .gitignore already exists, whether to append contents (TRUE) or overwrite (FALSE).
If left blank will prompt the user within R Studio.}
}
\value{
Adds .gitignore file to directory, or appends content to existing file.
}
\description{
Add PHS template .gitignore file to chosen directory.
}
\examples{
\dontrun{
add_gitignore()
}
}
