% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ignore_file.R
\name{add_ignore_file}
\alias{add_ignore_file}
\title{add_ignore_file}
\usage{
add_ignore_file()
}
\value{
Adds .gitignore file to directory, or appends content to existing file.
}
\description{
Add PHS template .gitignore file to chosen directory. This function is intended for use in R Studio.
}
\examples{
\dontrun{
add_ignore_file()
}
}
