% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stylecss.R
\name{add_stylecss}
\alias{add_stylecss}
\title{add_stylecss}
\usage{
add_stylecss(
  path = rstudioapi::selectDirectory(caption = "Select folder to add style.css")
)
}
\arguments{
\item{path}{String: path to add style.css file. If left blank, will prompt the user within R Studio}
}
\value{
Adds style.css file to directory.
}
\description{
Add PHS template style.css file to chosen directory.
}
\examples{
\dontrun{
add_stylecss()
}
}
