% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_report.R
\name{compile_report}
\alias{compile_report}
\title{compile_report}
\usage{
compile_report(
  rmd_filename = list.files(pattern = "\\\\.Rmd$")[1],
  cover_filename = "Cover_Page.docx",
  title = "My Title",
  subtitle = "My Subtitle",
  date = "DD Month YYYY",
  filename_out = "Report_and_Cover.docx",
  auto_open = TRUE
)
}
\arguments{
\item{rmd_filename}{Report filename.}

\item{cover_filename}{Cover page filename.}

\item{title}{Title for cover page.}

\item{subtitle}{Subtitle for cover page.}

\item{date}{Date of Publication.}

\item{filename_out}{Output filename for compiled report.}

\item{auto_open}{Automatically open the report after it is compiled?}
}
\value{
Report with table of contents and front cover included in docx format.
}
\description{
Compile report with table of contents and front cover included.
}
\examples{
\dontrun{
compile_report()
}
}
