% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_script.R
\name{new_script}
\alias{new_script}
\title{new_script}
\usage{
new_script()
}
\value{
New R script created according to the PHS R project structure.
}
\description{
Create new R script according to the PHS template. This function is meant to be used within RStudio by clicking on the Addins dropdown menu.
}
\examples{
\dontrun{
new_script()
}
}
