% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsproject.R
\name{phsproject}
\alias{phsproject}
\title{phsproject}
\usage{
phsproject(
  path,
  author,
  n_scripts = 1,
  git = FALSE,
  renv = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{path}{Filepath for the project.}

\item{author}{Name of the main author for the project.}

\item{n_scripts}{Number of code files to start the project with.}

\item{git}{Initialise the project with Git.}

\item{renv}{Initialise the project with package management using renv.}

\item{overwrite}{Logical: Whether to overwrite directory at existing path when creating directory.}
}
\value{
New project created according to the PHS R project structure.
}
\description{
Create new projects according to the PHS R project structure. This function is meant to be used within RStudio by going to the File menu, then New Project.
}
\examples{
\dontrun{
phsproject(path = file.path(getwd(), "testproj"), author = "A Person", n_scripts = 1)
}
}
