% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{update_metadata}
\usage{
update_metadata()
}
\value{
Updated metadata in the R script.
}
\description{
Updates metadata in the PHS R script template for existing script. This function is meant to be used within RStudio by clicking on the Addins dropdown menu.
}
\examples{
\dontrun{
update_metadata()
}
}
