% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{add_gitignore}
\usage{
add_gitignore()
}
\value{
Adds .gitignore file to directory, or appends content to existing file.
}
\description{
Add PHS template .gitignore file to chosen directory. This function is intended for use in R Studio.
}
\examples{
\dontrun{
add_gitignore()
}
}
