% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phs_html.R
\name{create_phs_html}
\alias{create_phs_html}
\title{create_phs_html}
\usage{
create_phs_html(file_name = NULL, ext_name = "phs-html-quarto")
}
\arguments{
\item{file_name}{Character string of name given to output `.qmd` file.}

\item{ext_name}{Character string of extension folder name (default = "phs-html-quarto")}
}
\value{
Quarto document for html with PHS formatting.
}
\description{
Creates template `.qmd` file for a html report with PHS formatting.
}
\examples{
\dontrun{
create_phs_html()
}
}
