% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiproject.R
\name{phiproject}
\alias{phiproject}
\title{phiproject}
\usage{
phiproject(path, author, n_scripts = 1, git = FALSE, renv = FALSE)
}
\arguments{
\item{path}{Filepath for the project.}

\item{author}{Name of the main author for the project.}

\item{n_scripts}{Number of code files to start the project with.}

\item{git}{Initialise the project with Git.}

\item{renv}{Initialise the project with package management using renv.}
}
\value{
New project created according to the PHI R project structure.
}
\description{
Create new projects according to the PHI R project structure. This function is meant to be used within RStudio by going to the File menu, then New Project.
}
\examples{
\dontrun{
phiproject(path = file.path(getwd(), "testproj"), author = "A Person", n_scripts = 1, git = FALSE)
}
}
