% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiproject.R
\name{phiproject}
\alias{phiproject}
\title{phiproject}
\usage{
phiproject(path, author)
}
\arguments{
\item{path}{Filepath for the project.}

\item{author}{Name of the main author for the project.}
}
\value{
New project created according to the PHI R project structure.
}
\description{
Create new projects according to the PHI R project structure
}
\examples{
phiproject(path = file.path(getwd(), "testproj"), author = "A Person")
}
