% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsshinyapp.R
\name{phsshinyapp}
\alias{phsshinyapp}
\title{phsshinyapp}
\usage{
phsshinyapp(
  path,
  author,
  app_name = "WRITE APP NAME HERE",
  git = FALSE,
  renv = FALSE
)
}
\arguments{
\item{path}{String: Filepath for the project.}

\item{author}{String: Name of the main author for the project.}

\item{app_name}{String: name of application.}

\item{git}{Logical: Initialise the project with Git.}

\item{renv}{Logical: Initialise the project with package management using renv.}
}
\value{
New project created according to the PHS R project structure.
}
\description{
Create new shiny app according to the PHS R project structure.
This function is meant to be used within RStudio by going to the File menu, then New Project.
}
\examples{
\dontrun{
phsshinyapp(path = file.path(getwd(), "testproj"), author = "A Person", n_scripts = 1)
}
}
