% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phstemplates.R
\docType{package}
\name{phstemplates}
\alias{phstemplates}
\title{\code{phstemplates} package}
\description{
Standard Templates for use in PHS.
}
\details{
See the README on
\href{https://github.com/Public-Health-Scotland/phstemplates#readme}{GitHub}.
}
